backend/requirements.txt
fastapi
uvicorn
pandas
numpy
prophet
statsmodels
pmdarima
google-generativeai
reportlab
matplotlib
python-multipart
python-dotenv
scikit-learn
pytest
httpx
